+++ tests/functional/resolvers.ts
@@ -1613,6 +1613,25 @@ describe("Resolvers", () => {
       expect(data!.sampleQuerySync.sampleFieldSync).toEqual("sampleFieldSync");
     });
 
+    it("should generate the schema when schema is incorrect but `skipCheck` is set to true", async () => {
+      getMetadataStorage().clear();
+
+      @Resolver()
+      class SampleResolver {
+        @Mutation()
+        sampleMutation(): string {
+          return "sampleMutation";
+        }
+      }
+
+      const schema = await buildSchema({
+        resolvers: [SampleResolver],
+        skipCheck: true,
+      });
+
+      expect(schema).toBeDefined();
+    });
+
     it("should throw errors when no resolvers provided", async () => {
       getMetadataStorage().clear();
       expect.assertions(2);