+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 - add postinstall script for printing info on console about supporting the project
 - add support for setting default nullability for fields and return types (#297)
+- add `skipCheck` option in `buildSchema` to disable checking the correctness of a schema
 ### Fixes
 - fix generating plain resolvers for queries and mutations (compatibility with Apollo client state)